+++ asmar-backend/src/main/java/com/aston/ecommerce/asmar/dao/CommandLineRepository.java
@@ -2,13 +2,11 @@ package com.aston.ecommerce.asmar.dao;
 
 import com.aston.ecommerce.asmar.entity.CommandLine;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
-
 
 import java.util.List;
 
-public interface CommandLineRepository extends JpaRepository<CommandLine, Long> {
-    List<CommandLine> findAllByUserIdAndOrderIsNull(Long userId);
-    List<CommandLine> findByOrderId(Long orderId);
-    CommandLine findByProductIdAndOrderIsNull(Long productId);
+public interface CommandLineRepository extends JpaRepository<CommandLine, Integer> {
+    List<CommandLine> findAllByUserIdAndOrderIsNull(Integer userId);
+    List<CommandLine> findByOrderId(Integer orderId);
+    CommandLine findByProductIdAndOrderIsNull(Integer productId);
 }
